requirements.txt
# Tally FastAPI Database Loader
# Python Dependencies

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
aiosqlite==0.19.0

# HTTP Client (for Tally communication)
httpx==0.26.0

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.0

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# XML Processing
lxml==5.1.0

# Templates & Static Files
jinja2==3.1.3
python-multipart==0.0.6

# Logging
loguru==0.7.2

# WebSocket support (for real-time logs)
websockets==12.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3

# Development
black==23.12.1
isort==5.13.2
